Remove --account-name parameter when it's required
This makes commands that require account-name easier to write, i.e.

```
hal config provider kubernetes add-account my-k8s-account --context my-context
```

vs.

```
hal config provider kubernetes add-account --account-name my-k8s-account --context my-context
```
