+++ halyard-cli/src/main/java/com/netflix/spinnaker/halyard/cli/command/v1/providers/AbstractDeleteAccountCommand.java
@@ -16,15 +16,11 @@
 
 package com.netflix.spinnaker.halyard.cli.command.v1.providers;
 
-import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
 import com.netflix.spinnaker.halyard.cli.command.v1.NestableCommand;
 import com.netflix.spinnaker.halyard.cli.services.v1.Daemon;
 import com.netflix.spinnaker.halyard.cli.ui.v1.AnsiUi;
-import com.netflix.spinnaker.halyard.config.model.v1.node.Account;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -33,7 +29,7 @@ import lombok.Getter;
  * Delete a specific PROVIDER account
  */
 @Parameters()
-public abstract class AbstractDeleteAccountCommand extends AbstractProviderCommand {
+public abstract class AbstractDeleteAccountCommand extends AbstractHasAccountCommand {
   @Getter(AccessLevel.PROTECTED)
   private Map<String, NestableCommand> subcommands = new HashMap<>();
 
@@ -41,26 +37,7 @@ public abstract class AbstractDeleteAccountCommand extends AbstractProviderComma
   private String commandName = "delete-account";
 
   public String getDescription() {
-    return "Delete a specific " + getProviderName() + " account by name";
-  }
-
-  @Parameter(description = "The name of the account to delete", arity = 1)
-  List<String> accounts = new ArrayList<>();
-
-  @Override
-  public String getMainParameter() {
-    return "account";
-  }
-
-  public String getAccountName() {
-    switch (accounts.size()) {
-      case 0:
-        throw new IllegalArgumentException("No account name supplied");
-      case 1:
-        return accounts.get(0);
-      default:
-        throw new IllegalArgumentException("More than one account supplied");
-    }
+    return "Delete a specific " + getProviderName() + " account by name.";
   }
 
   @Override
